requirements.txt
asgiref==3.6.0
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
dj-rest-auth==3.0.0
django-allauth==0.52.0
gunicorn==20.1.0
numpy==1.24.2
Pillow==9.4.0
psycopg2==2.9.5
pytz==2022.7.1
sqlparse==0.4.3
uvicorn==0.20.0
